+++ .github/workflows/iroha2-dev.yml
@@ -3,24 +3,25 @@ name: Iroha 2 dev branch workflow
 on:
   push:
     branches: [iroha2-dev]
+
 env:
-  CACHE_PATHS: |
-    ~/.cargo/registry
-    ~/.cargo/git
-    ~/.rustup
-    target
-    iroha/target
-    iroha_client/target
+  CARGO_TERM_COLOR: always
+
 jobs:
   archive-artifacts:
     runs-on: [self-hosted, Linux]
-    container: rust:1.54-buster
+    container: rust:1.55-buster
     steps:
       - uses: actions/checkout@v2
       - uses: actions/cache@v2
         with:
-          path: ${{ env.CACHE_PATHS }}
-          key: ${{ hashFiles('Cargo.toml') }}
+          path: |
+              ${{ env.CARGO_HOME }}
+              target/
+          key: iroha2-rust-archive-${{ hashFiles('Cargo.lock') }}
+          restore-keys: |
+              iroha2-rust-archive-
+
       - name: Install dependencies
         run: |
           apt-get update
@@ -49,7 +50,7 @@ jobs:
 
   deploy:
     runs-on: [self-hosted, Linux]
-    container: rust:1.54-buster
+    container: rust:1.55-buster
     steps:
       - uses: actions/checkout@v2
       - name: Check Dockerhub credentials
@@ -88,7 +89,7 @@ jobs:
   # 2  Coverage bot can have results from `iroha2-dev` to report accurate coverage changes.
   coverage:
     runs-on: [self-hosted, Linux]
-    container: rust:1.54-buster
+    container: rust:1.55-buster
     steps:
       - uses: actions/checkout@v2
       - name: Install dependencies
@@ -130,8 +131,11 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions/cache@v2
         with:
-          path: ${{ env.CACHE_PATHS }}
-          key: ${{ hashFiles('Cargo.toml') }}
+          path: ${{ env.RUST_CACHE_PATHS }}
+          key: iroha2-rust-publish-artifacts-${{ hashFiles('Cargo.lock') }}
+          restore-keys: |
+              iroha2-rust-publish-artifacts-
+
       - name: Run schema generation
         run: |
           mkdir -p target/schema
