+++ iroha_client/tests/merge.rs
@@ -1,0 +1,1 @@
+mod tests;
