+++ iroha_client/tests/tests/add_asset.rs
