+++ iroha_client/tests/tests/add_domain.rs
