+++ iroha_client/tests/tests/events.rs
