+++ iroha_client/tests/tests/restart_peer.rs
